Improve envelope card (#119)
* Create EnvelopList file

* Improve envelope card

* Improve translations

* Move EnvelopeCard to its own component and remove unnecessary comments

---------
