+++ src/components/Process/Detail.tsx
@@ -52,6 +52,7 @@ import { useElectionKeys, useElectionVotesList } from '~queries/processes'
 import { ucfirst } from '~utils/strings'
 import { RouteParamsTabs } from '~components/Layout/RouteParamsTabs'
 import { NoResultsError } from '~components/Layout/ContentError'
+import { PaginatedEnvelopeList } from '~components/Envelope/EnvelopeList'
 
 const Detail = () => {
   const { election } = useElection()
@@ -171,7 +172,7 @@ const Detail = () => {
             <ElectionResults />
           </TabPanel>
           <TabPanel>
-            <EnvelopeExplorer />
+            <PaginatedEnvelopeList />
           </TabPanel>
           <TabPanel>
             <RawContentBox obj={raw} />
@@ -215,105 +216,6 @@ const ElectionKeys = ({ electionId }: { electionId: string }) => {
         </GridItem>
       </Grid>
     </Flex>
-  )
-}
-
-const EnvelopeExplorer = () => {
-  const { election: e } = useElection()
-  const election = e as PublishedElection
-  const { t } = useTranslation()
-
-  if (!election || election.voteCount === 0) {
-    return <NoResultsError msg={t('election.no_votes_yet', { defaultValue: 'No votes yet!' })} />
-  }
-
-  return (
-    <PaginationProvider>
-      <EnvelopeList />
-    </PaginationProvider>
-  )
-}
-
-const EnvelopeList = () => {
-  const { page } = usePagination()
-  const { election: e } = useElection()
-  const election = e as PublishedElection
-
-  const { data: envelopes, isLoading } = useElectionVotesList({
-    params: {
-      electionId: election?.id,
-      page: page,
-    },
-    enabled: !!election?.id,
-  })
-
-  if (isLoading) {
-    return <LoadingCards />
-  }
-
-  if (!envelopes) {
-    // todo(kon): add error handling
-    return null
-  }
-
-  return (
-    <Flex direction={'column'} gap={4}>
-      <Grid
-        templateColumns={{ base: 'repeat(1, 1fr)', sm: 'repeat(2, 1fr)', md: 'repeat(3, 1fr)', lg: 'repeat(5, 1fr)' }}
-        gap={4}
-      >
-        {envelopes?.votes.map((envelope, i) => {
-          return <EnvelopeCard key={i} envelope={envelope} count={page * 10 + i + 1} />
-        })}
-      </Grid>
-      <Pagination pagination={envelopes.pagination} />
-    </Flex>
-  )
-}
-
-const EnvelopeCard = ({ envelope, count }: { envelope: VoteSummary; count: number }) => {
-  return (
-    <Card maxW='xs'>
-      <CardHeader>
-        <Flex justify={'space-between'}>
-          <Text fontWeight={'bold'}>
-            <Trans i18nKey={'envelopes.envelope_number'} num={count}>
-              Envelope nº {{ num: count }}
-            </Trans>
-          </Text>
-          <Box>
-            <Icon color={'lightText'} as={BiEnvelope} />
-          </Box>
-        </Flex>
-      </CardHeader>
-      <CardBody>
-        <Flex direction={'column'}>
-          <Link
-            as={RouterLink}
-            to={generatePath(RoutePath.Block, { height: envelope.blockHeight.toString(), tab: null, page: null })}
-          >
-            <Trans i18nKey={'envelopes.block'} height={envelope.blockHeight}>
-              Block {{ height: envelope.blockHeight }}
-            </Trans>
-          </Link>
-          <Link
-            as={RouterLink}
-            to={generatePath(RoutePath.Transaction, {
-              block: envelope.blockHeight.toString(),
-              index: envelope.transactionIndex.toString(),
-              tab: null,
-            })}
-          >
-            <Trans i18nKey={'envelopes.tx_number'} transactionIndex={envelope.transactionIndex}>
-              Transaction: {{ transactionIndex: envelope.transactionIndex }}
-            </Trans>
-          </Link>
-          <Link as={RouterLink} to={generatePath(RoutePath.Envelope, { verifier: envelope.voteID, tab: null })}>
-            <Trans i18nKey={'envelopes.details'}>Details</Trans>
-          </Link>
-        </Flex>
-      </CardBody>
-    </Card>
   )
 }
